wip refactor firearms data structure
