+++ src/data/firearms/aimTimes.js
@@ -1,0 +1,53 @@
+// Rifles
+
+// M1949-56, SVT 40, MAS 36, Kar 98k
+export const averageRifle11Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
+  mod: [-23, -12, -9, -7, -6, -5, -4, -3, -2, -1, 0],
+};
+
+// HK G3, L1A1
+export const belowAverageverageRifle11Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
+  mod: [-24, -14, -9, -8, -6, -5, -4, -3, -2, -1, 0],
+};
+
+// M16, M16A1
+export const aboveAverageRifle11Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
+  mod: [-22, -12, -9, -7, -6, -5, -4, -3, -2, -1, 0],
+};
+
+// M1 Carbine, M2 Carbine
+export const improvedRifle11Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
+  mod: [-21, -11, -9, -7, -6, -4, -3, -2, -2, -1, 0],
+};
+
+// AK47, AKM, AK74, FAMAS
+export const averageRifle9Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8, 9],
+  mod: [-23, -12, -9, -7, -6, -4, -3, -2, -1],
+};
+
+// SMGs
+
+// Carl Gustaf M45, Sten Mk2
+export const averageSMG8Aims = {
+  ac: [1, 2, 3, 4, 5, 6, 7, 8],
+  mod: [-22, -12, -9, -7, -6, -5, -4, -3],
+};
+
+// Pistols
+
+// FN Mk1, Lebel M1892, Tokarev TT33,
+export const averagePistol6Aims = {
+  ac: [1, 2, 3, 4, 5, 6],
+  mod: [-17, -11, -10, -9, -8, -7],
+};
+
+// M1911A1, MAB PA15
+export const belowAveragePistol6Aims = {
+  ac: [1, 2, 3, 4, 5, 6],
+  mod: [-18, -11, -10, -9, -8, -7],
+};
